+++ MOVEit.MIT/Public/Get-MITFileContent.ps1
@@ -1,7 +1,7 @@
-function Read-MITFile {
+function Get-MITFileContent {
     <#
     .SYNOPSIS
-        Read (download) a MOVEit Transfer file/attachment
+        Get (download) a MOVEit Transfer file/attachment
     .LINK
         Download file
         https://docs.ipswitch.com/MOVEit/Transfer2021/Api/Rest/#operation/GETapi/v1/files/{Id}/download-1.0